only get date from mip sampleinfo file

+++ cg/cli/clean.py
@@ -50,38 +50,40 @@ def beacon(context: click.Context, item_type, item_id):
 
 @clean.command()
 @click.option("-y", "--yes", is_flag=True, help="skip confirmation")
+@click.argument("case_id")
 @click.argument("sample_info", type=click.File("r"))
 @click.pass_context
-def mip(context, yes, sample_info):
+def mip(context, yes, case_id, sample_info):
     """Remove analysis output."""
-    raw_data = ruamel.yaml.safe_load(sample_info)
-    data = context.obj["tb"].parse_sampleinfo(raw_data)
 
-    family = data["family"]
-    family_obj = context.obj["db"].family(family)
-    if family_obj is None:
-        LOG.error(f"{family}: family not found")
+    raw_data = ruamel.yaml.safe_load(sample_info)
+    date = context.obj["tb"].parse_sample_date(raw_data)
+    case_obj = context.obj["db"].family(case_id)
+
+    if case_obj is None:
+        LOG.error(f"{case_id}: family not found")
         context.abort()
 
-    analysis_obj = context.obj["db"].analysis(family_obj, data["date"])
+    analysis_obj = context.obj["db"].analysis(case_obj, date)
     if analysis_obj is None:
-        LOG.error(f"{family} - {data['date']}: analysis not found")
+        LOG.error(f"{case_id} - {date}: analysis not found")
         context.abort()
 
     try:
-        context.obj["tb"].delete_analysis(family, data["date"], yes=yes)
+        context.obj["tb"].delete_analysis(case_id, date, yes=yes)
     except ValueError as error:
-        LOG.error(f"{family}: {error.args[0]}")
+        LOG.error(f"{case_id}: {error.args[0]}")
         context.abort()
 
 
 @clean.command()
 @click.argument("bundle")
 @click.option("-y", "--yes", is_flag=True, help="skip checks")
+@click.option("-d", "--dry-run", is_flag=True, help="show files that would be cleaned")
 @click.pass_context
-def scout(context, bundle, yes):
+def scout(context, bundle, yes, dry_run):
     files = []
-    for tag in ["bam", "bai", "bam-index"]:
+    for tag in ["bam", "bai", "bam-index", "cram", "crai", "cram-index"]:
         files.extend(context.obj["hk"].get_files(bundle=bundle, tags=[tag]))
     for file_obj in files:
         if file_obj.is_included:
@@ -92,17 +94,22 @@ def scout(context, bundle, yes):
         if yes or click.confirm(question):
             file_name = file_obj.full_path
             if file_obj.is_included and Path(file_name).exists():
-                Path(file_name).unlink()
+                if not dry_run:
+                    Path(file_name).unlink()
 
-            file_obj.delete()
-            context.obj["hk"].commit()
-            click.echo(f"{file_name} deleted")
+            if not dry_run:
+                file_obj.delete()
+                context.obj["hk"].commit()
+                click.echo(f"{file_name} deleted")
 
 
 @clean.command()
 @click.option("-y", "--yes", is_flag=True, help="skip checks")
+@click.option(
+    "-d", "--dry-run", is_flag=True, help="Shows cases and files that would be cleaned"
+)
 @click.pass_context
-def scoutauto(context, yes):
+def scoutauto(context, yes, dry_run):
     """Automatically clean up solved and archived scout cases"""
     bundles = []
     for status in "archived", "solved":
@@ -116,7 +123,7 @@ def scoutauto(context, yes):
         LOG.info(f"{cases_added} cases marked for bam removal :)")
 
     for bundle in bundles:
-        context.invoke(scout, bundle=bundle, yes=yes)
+        context.invoke(scout, bundle=bundle, yes=yes, dry_run=dry_run)
 
 
 @clean.command()
@@ -128,29 +135,31 @@ def mipauto(context: click.Context, before_str: str, yes: bool = False):
     before = parse_date(before_str)
     old_analyses = context.obj["db"].analyses(before=before)
     for status_analysis in old_analyses:
-        family_id = status_analysis.family.internal_id
-        LOG.debug(f"{family_id}: clean up analysis output")
+        case_id = status_analysis.family.internal_id
+        LOG.debug(f"{case_id}: clean up analysis output")
         tb_analysis = context.obj["tb"].find_analysis(
-            family=family_id, started_at=status_analysis.started_at, status="completed"
+            family=case_id, started_at=status_analysis.started_at, status="completed"
         )
 
         if tb_analysis is None:
-            LOG.warning(f"{family_id}: analysis not found in Trailblazer")
+            LOG.warning(f"{case_id}: analysis not found in Trailblazer")
             continue
         elif tb_analysis.is_deleted:
-            LOG.warning(f"{family_id}: analysis already deleted")
+            LOG.warning(f"{case_id}: analysis already deleted")
             continue
-        elif context.obj["tb"].analyses(family=family_id, temp=True).count() > 0:
-            LOG.warning(f"{family_id}: family already re-started")
+        elif context.obj["tb"].analyses(family=case_id, temp=True).count() > 0:
+            LOG.warning(f"{case_id}: family already re-started")
             continue
 
         try:
             sampleinfo_path = context.obj["tb"].get_sampleinfo(tb_analysis)
-            LOG.info(f"{family_id}: cleaning MIP output")
+            LOG.info(f"{case_id}: cleaning MIP output")
             with open(sampleinfo_path, "r") as sampleinfo_file:
-                context.invoke(mip, yes=yes, sample_info=sampleinfo_file)
-        except FileNotFoundError as err:
+                context.invoke(
+                    mip, yes=yes, case_id=case_id, sample_info=sampleinfo_file
+                )
+        except FileNotFoundError:
             LOG.error(
-                f"{family_id}: sample_info file not found, please mark the analysis as deleted in "
+                f"{case_id}: sample_info file not found, please mark the analysis as deleted in "
                 f"the analysis table in trailblazer."
             )
